+++ packages/xization/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v1.3.6 - 2026-05-18
+
 ### Added
 
 - **`BackendDocumentSource(file_filter=)`** — optional keyword-only